backend/requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
faiss-cpu
PyMuPDF
python-multipart
ollama
